requirements.txt
streamlit
numpy
opencv-python-headless
matplotlib
scikit-learn
pillow
scipy